document sidebar framework, files, activity
