+++ app/app/styles/view/common.scss
@@ -3,15 +3,13 @@
 
 	> .empty-state {
 		margin: 30px 0;
-		font-size: 2rem;
+		font-size: 1.2rem;
 		color: $color-gray;
-		font-family: 'open_sanslight';
 	}
 
 	> .normal-state {
 		margin: 10px;
-		font-size: 2rem;
+		font-size: 1.2rem;
 		color: $color-gray;
-		font-family: 'open_sanslight';
 	}
 }
