--- app/app/styles/view/document/activity.scss
@@ -1,139 +1,0 @@
-.document-activity {
-	> .metrics {
-		list-style-type: none;
-		margin: 0 0 30px;
-		padding: 0;
-		width: 100%;
-		text-align: center;
-
-		> .metric {
-			padding: 0 30px;
-			display: inline-block;
-			text-align: center;
-
-			.label {
-				padding: 0 0 10px;
-				font-family: $font-regular;
-				font-size: 0.9rem;
-				color: $color-gray;
-			}
-
-			.number {
-				font-family: $font-light;
-				font-size: 2rem;
-				color: $color-off-black;
-			}
-		}
-	}
-
-	> .items {
-		list-style-type: none;
-		margin: 0;
-		padding: 0;
-		white-space: nowrap;
-
-		> .item {
-			padding: 20px 0;
-			width: 100%;
-			border-bottom: 1px solid $color-border;
-
-			&:last-of-type {
-				border-bottom: none !important;
-			}
-
-			> .avatar-box {
-				display: inline-block;
-				margin: 3px 10px 0 0;
-				width: 5%;
-			}
-
-			> .name {
-				display: inline-block;
-				font-size: 1rem;
-				font-family: $font-light;
-				color: $color-off-black;
-				width: 25%;
-			}
-
-			> .detail {
-				display: inline-block;
-				font-size: 1rem;
-				font-family: $font-regular;
-				color: $color-off-black;
-				width: 45%;
-
-				.viewed {
-					color: $color-goldy;
-				}
-
-				.added {
-					color: $color-green;
-				}
-
-				.changed {
-					color: $color-blue;
-				}
-
-				.deleted {
-					color: $color-red;
-				}
-			}
-
-			> .date {
-				display: inline-block;
-				font-family: $font-light;
-				font-size: 1rem;
-				float: right;
-				width: 15%;
-				padding-top: 10px;
-			}
-		}
-	}
-}
-
-.meta-editors {
-	position: relative;
-
-	> .items {
-		list-style-type: none;
-		margin: 0;
-		padding: 0;
-		white-space: nowrap;
-
-		> .item {
-			margin: 15px 0;
-			overflow-x: hidden;
-			white-space: nowrap;
-			text-overflow: ellipsis;
-			width: 90%;
-
-			.avatar-box {
-				display: inline-block;
-				margin: 3px 10px 0 0;
-				vertical-align: top;
-			}
-
-			.detail {
-				display: inline-block;
-
-				.name {
-					font-size: 1rem;
-					color: $color-off-black;
-				}
-
-				.changed {
-					font-size: 0.9rem;
-				}
-
-				.deleted {
-					font-size: 0.9rem;
-					color: $color-red;
-				}
-
-				.date {
-					font-size: 0.8rem;
-				}
-			}
-		}
-	}
-}
