+++ app/app/styles/view/document/all.scss
@@ -1,11 +1,10 @@
-@import "activity.scss";
 @import "content-linker.scss";
-@import "files.scss";
 @import "history.scss";
 @import "inline-editor.scss";
 @import "layout.scss";
 @import "section-editor.scss";
-@import "sidebar.scss";
-@import "toolbar.scss";
+@import "sidebar-view-attachments.scss";
+@import "sidebar-view-activity.scss";
+@import "sidebar-zone.scss";
 @import "view.scss";
 @import "wysiwyg.scss";
