+++ app/app/styles/view/document/sidebar-view-activity.scss
@@ -1,0 +1,52 @@
+.document-sidebar-view-activity {
+	> .items {
+		list-style-type: none;
+		margin: 0;
+		padding: 0;
+		white-space: nowrap;
+
+		> .item {
+			margin: 0;
+			padding: 10px 0;
+			width: 100%;
+
+			> .avatar-box {
+				display: inline-block;
+				margin: 0 10px 0 0;
+			}
+
+			> .name {
+				display: inline-block;
+				font-size: 0.9rem;
+				color: $color-gray;
+				width: 200px;
+				@extend .truncate;			
+			}
+
+			> .detail {
+				display: block;
+				font-size: 0.9rem;
+				color: $color-off-black;
+				margin-left: 50px;
+				width: 200px;
+				@extend .truncate;			
+
+				.viewed {
+					color: $color-goldy;
+				}
+
+				.added {
+					color: $color-green;
+				}
+
+				.changed {
+					color: $color-blue;
+				}
+
+				.deleted {
+					color: $color-red;
+				}
+			}
+		}
+	}
+}
