+++ app/app/styles/view/document/sidebar-view-attachments.scss
@@ -1,0 +1,144 @@
+.document-sidebar-view-attachments {
+	margin: 0;
+
+	> .upload-document-files {
+		width: 100%;
+		padding: 20px;
+		margin-bottom: 20px;
+		text-align: center;
+		color: $color-gray;
+		border: 1px solid $color-stroke;
+		cursor: pointer;
+		font-size: 0.9rem;
+		line-height: 1.7rem;
+		@include ease-in();
+
+		&:hover {
+			border-color: $color-link;
+			color: $color-link;
+		}
+
+		 > .dz-preview,
+		.dz-processing {
+			display: none !important;
+		}
+	}
+
+	> .list {
+		margin: 0 0 50px;
+		padding: 7px 0;
+
+		> .item {
+			color: $color-off-black;
+			margin: 0;
+			padding: 10px 0;
+			font-size: 0.9rem;
+			list-style-type: none;
+
+			&:last-of-type {
+				border-bottom: none !important;
+			}
+
+			> .icon {
+				margin-right: 10px;
+			}
+
+			> a {
+				@extend .truncate;
+				width: 200px;
+				color: $color-gray;
+
+				&:hover {
+					color: $color-link;
+				}
+
+				> .file {
+					@extend .truncate;
+					display: inline-block;
+					font-size: 0.9rem;
+					width: 200px;
+				}
+			}
+
+			> .action {
+				float: right;
+				margin-top: -2px;
+				margin-right: 5px;
+				@extend .cursor-pointer;
+				@extend .transition-all;
+				display: none;
+				color: $color-gray;
+			}
+
+			&:hover {
+				.action {
+					display: inline-block;
+				}
+			}
+		}
+	}
+}
+
+.delete-attachment-dialog,
+.delete-page-dialog {
+	display: none;
+}
+
+
+
+// 	.document-structure {
+// 		> .toc-controls {
+// 			margin: 0;
+// 			color: $color-gray;
+
+// 			> .round-button-mono {
+// 				color: $color-green;
+// 				border-color: $color-green;
+
+// 				> .material-icons {
+// 					color: $color-green;
+// 				}
+// 			}
+
+// 			> .disabled {
+// 				@extend .cursor-not-allowed;
+// 				color: $color-stroke;
+// 				border-color: $color-stroke;
+
+// 				> .material-icons {
+// 					color: $color-stroke;
+// 				}
+// 			}
+// 		}
+
+// 		.entries {
+// 			padding: 0;
+// 			list-style: none;
+// 			font-size: 13px;
+// 			overflow-x: hidden;
+// 			list-style-type: none;
+// 			margin: 20px 0 0;
+// 			font-family: $font-semibold;
+
+// 			.item {
+// 				padding: 4px 0;
+// 				text-overflow: ellipsis;
+// 				word-wrap: break-word;
+// 				white-space: nowrap;
+// 				overflow: hidden;
+
+// 				> .link {
+// 					color: $color-gray;
+
+// 					&:hover {
+// 						color: $color-link;
+// 					}
+// 				}
+
+// 				> .selected {
+// 					color: $color-link;
+// 					font-weight: bold;
+// 				}
+// 			}
+// 		}
+// 	}
