--- app/app/styles/view/document/toolbar.scss
@@ -1,114 +1,0 @@
-.document-toolbar {
-	position: relative;
-	margin: -30px 0 40px;
-	height: 60px;
-	padding: 5px 30px 0;
-	background-color: $color-sidebar;
-	@include border-radius-bottom-right(5px);
-	@include border-radius-bottom-left(5px);
-	@extend .no-select;
-
-	> .tabs {
-		width: 70%;
-		height: 50px;
-		margin-top: 15px;
-		display: inline-block;
-
-		> .tab {
-			list-style-type: none;
-			display: inline-block;
-			margin: 0 20px 0 0;
-			padding: 0;
-			color: $color-gray;
-			font-family: $font-semibold;
-			cursor: pointer;
-
-			&:hover {
-				color: $color-link;
-			}
-
-			> a {
-				color: $color-gray;
-				@include ease-in();
-
-				&:hover {
-					color: $color-link;
-				}
-			}
-
-			> .active {
-				color: $color-link;
-
-				> .add-tab {
-					> i {
-						color: $color-link;
-					}
-				}
-			}
-
-			.add-tab {
-				display: inline-block;
-				vertical-align: text-top;
-				@include ease-in();
-
-				> i {
-					font-size: 1.5rem;
-					color: $color-gray;
-
-					&:hover {
-						color: $color-link;
-					}
-				}
-			}
-		}
-	}
-
-	> .options {
-		width: 30%;
-		height: 50px;
-		margin-top: 15px;
-		display: inline-block;
-		text-align: right;
-		float: right;
-
-		> .option {
-			list-style-type: none;
-			display: inline-block;
-			margin: 0 0 0 20px;
-			padding: 0;
-			color: $color-gray;
-			cursor: pointer;
-			@include ease-in();
-
-			&:hover {
-				color: $color-link;
-			}
-
-			> .pinned {
-				color: $color-primary;
-			}
-		}
-
-		> a  {
-			display: inline-block;
-
-			> .option {
-				list-style-type: none;
-				display: inline-block;
-				margin: 0 0 0 20px;
-				padding: 0;
-				color: $color-gray;
-				cursor: pointer;
-				@include ease-in();
-
-				&:hover {
-					color: $color-link;
-				}
-
-				> .pinned {
-					color: $color-primary;
-				}
-			}
-		}
-	}
-}
