+++ app/app/styles/widget/widget-avatar.scss
@@ -3,7 +3,7 @@
     background-color: $color-gray;
     @include border-radius(20px);
 	@include ease-in();
-    padding: 7px 0 0 0;
+    padding: 10px 0 0 0;
     letter-spacing: 1px;
     text-align: center;
     height: 35px;
