+++ app/app/templates/components/document/content-linker.hbs
