--- app/app/templates/components/document/document-activity.hbs
@@ -1,41 +1,0 @@
-<div class="document-activity">
-	<ul class="metrics">
-		<li class="metric">
-			<div class="label">created</div>
-			<div class="number">{{time-ago document.created}}</div>
-		</li>
-		<li class="metric">
-			<div class="label">viewers</div>
-			<div class="number">{{activity.viewers.length}}</div>
-		</li>
-		<li class="metric">
-			<div class="label">actions</div>
-			<div class="number">{{activity.editors.length}}</div>
-		</li>
-	</ul>
-    <ul class="items">
-        {{#each sortedItems as |e|}}
-            <li class="item">
-                <div class="avatar-box">
-                    <div class="avatar">{{user-initials e.item.firstname e.item.lastname}}</div>
-                </div>
-				<div class="name">{{e.item.firstname}} {{e.item.lastname}}</div>
-                <div class="detail">
-                    {{#if e.item.deleted}}
-                        <div class="deleted">{{e.item.changeLabel}}</div>
-						{{/if}}
-					{{#if e.item.changed}}
-                        <div class="changed">{{e.item.changeLabel}}</div>
-                    {{/if}}
-					{{#if e.item.added}}
-                        <div class="added">{{e.item.changeLabel}}</div>
-                    {{/if}}
-					{{#if e.item.viewed}}
-                        <div class="viewed">{{e.item.changeLabel}}</div>
-                    {{/if}}
-                </div>
-				<div class="date">{{time-ago e.date}}</div>
-            </li>
-        {{/each}}
-    </ul>
-</div>
