--- app/app/templates/components/document/document-toolbar.hbs
@@ -1,116 +1,0 @@
-<div class="document-toolbar hidden-xs hidden-sm">
-
-	<ul class="tabs">
-		<li class="tab">
-			{{#link-to 'document.index'}}Content{{/link-to}}
-		</li>
-		<li class="tab">
-			{{#link-to 'document.files'}}Files{{/link-to}}
-		</li>
-		{{#if session.authenticated}}
-			<li class="tab">
-				{{#link-to 'document.activity'}}Activity{{/link-to}}
-			</li>
-		{{/if}}
-		{{#each tabs as |tab|}}
-			<li class="tab">
-				{{#link-to 'document.section' tab.id (query-params mode='view')}}{{tab.title}}{{/link-to}}
-			</li>
-		{{/each}}
-		{{#if isEditor}}
-			<li class="tab">
-				{{#link-to 'document.wizard'}}
-					<div id="add-document-tab" class="add-tab" data-tooltip="Tab" data-tooltip-position="bottom center">
-						<i class="material-icons">add</i>
-					</div>
-				{{/link-to}}
-			</li>
-		{{/if}}
-	</ul>
-
-	<ul class="options">
-		{{#if session.authenticated}}
-			{{#if pinState.isPinned}}
-				<li class="option" {{action 'unpin'}}><i class="material-icons pinned">star</i></li>
-			{{else}}
-				<li class="option" id="pin-document-button"><i class="material-icons">star_border</i></li>
-			{{/if}}
-		{{/if}}
-		{{#if session.authenticated}}
-			{{#link-to 'document.history'}}
-				<li class="option"><i class="material-icons">history</i></li>
-			{{/link-to}}
-		{{/if}}
-		{{#if isEditor}}
-			<li class="option" id="save-template-button"><i class="material-icons">content_copy</i></li>
-			<li class="option" id="set-meta-button"><i class="material-icons">settings</i></li>
-		{{/if}}
-		<li class="option" id="document-toolbar-menu"><i class="material-icons">more_horiz</i></li>
-	</ul>
-
-	{{#dropdown-menu target="document-toolbar-menu" position="bottom right" open="click" onOpenCallback=(action 'onMenuOpen') onCloseCallback=(action 'onMenuOpen')}}
-		<ul class="menu">
-			{{#if isEditor}}
-				<li class="item" {{action 'changeLayout'}} >{{layoutLabel}}</li>
-				<li class="divider"></li>
-			{{/if}}
-
-			<li class="item" id="print-document-button" {{action 'printDocument'}}>Print</li>
-
-			{{#if isEditor}}
-				<li class="divider"></li>
-				<li class="item danger" id="delete-document-button">Delete</li>
-			{{/if}}
-		</ul>
-	{{/dropdown-menu}}
-
-	{{#if menuOpen}}
-		{{#dropdown-dialog target="delete-document-button" position="bottom right" button="Delete" color="flat-red" onAction=(action 'deleteDocument')}}
-			<p>Are you sure you want to delete this document?</p>
-			<p>There is no undo!</p>
-		{{/dropdown-dialog}}
-	{{/if}}
-
-	{{#if session.authenticated}}
-		{{#unless pinState.isPinned}}
-			{{#dropdown-dialog target="pin-document-button" position="bottom right" button="Pin" color="flat-green" onAction=(action 'pin') focusOn="pin-document-name" }}
-				<div class="input-control">
-					<label>Pin Document</label>
-					<div class="tip">A 3 or 4 character name</div>
-					{{input type='text' id="pin-document-name" value=pinState.newName}}
-				</div>
-			{{/dropdown-dialog}}
-		{{/unless}}
-	{{/if}}
-
-	{{#if isEditor}}
-		{{#dropdown-dialog target="save-template-button" position="bottom right" button="Save as Template" color="flat-green" onAction=(action 'saveTemplate') focusOn="new-template-name" }}
-			<div class="input-control">
-				<label>Name</label>
-				<div class="tip">Short name for this type of document</div>
-				{{input type='text' id="new-template-name" value=saveTemplate.name}}
-			</div>
-			<div class="input-control">
-				<label>Excerpt</label>
-				<div class="tip">Explain use case for this template</div>
-				{{textarea value=saveTemplate.description rows="3" id="new-template-desc"}}
-			</div>
-		{{/dropdown-dialog}}
-
-		{{#dropdown-dialog target="set-meta-button" position="bottom right" button="Save" color="flat-green" onAction=(action 'saveMeta') focusOn="meta-name" }}
-			<div class="input-control">
-				<label>Name</label>
-				<div class="tip">Short title for this document</div>
-				{{focus-input type='text' id="meta-name" value=document.name}}
-			</div>
-			<div class="input-control">
-				<label>Excerpt</label>
-				<div class="tip">Provide short summary of the document (max. 250)</div>
-				{{textarea value=document.excerpt rows="5" id="meta-excerpt"}}
-			</div>
-		{{/dropdown-dialog}}
-	{{/if}}
-
-</div>
-
-<div class="clearfix" />
