+++ app/app/templates/components/document/sidebar-view-activity.hbs
@@ -1,0 +1,29 @@
+<div class="document-sidebar-panel">
+	<div class="title">Activity</div>
+	<div class="document-sidebar-view-activity">
+		<ul class="items">
+			{{#each sortedItems as |e|}}
+				<li class="item">
+					<div class="avatar-box">
+						<div class="avatar">{{user-initials e.item.firstname e.item.lastname}}</div>
+					</div>
+					<div class="name">{{e.item.firstname}} {{e.item.lastname}}, {{time-ago e.date}}</div>
+					<div class="detail">
+						{{#if e.item.deleted}}
+							<div class="deleted">{{e.item.changeLabel}}</div>
+							{{/if}}
+						{{#if e.item.changed}}
+							<div class="changed">{{e.item.changeLabel}}</div>
+						{{/if}}
+						{{#if e.item.added}}
+							<div class="added">{{e.item.changeLabel}}</div>
+						{{/if}}
+						{{#if e.item.viewed}}
+							<div class="viewed">{{e.item.changeLabel}}</div>
+						{{/if}}
+					</div>
+				</li>
+			{{/each}}
+		</ul>
+	</div>
+</div>
