--- app/app/templates/components/layout/zone-document-sidebar.hbs
@@ -1,3 +1,0 @@
-<p>
-	sidebar
-</p>
